Fix all bugs, update JDA to 5.0.0-beta.1

+++ src/main/java/dev/cybo/tickets/commands/TicketOptionsCommand.java
@@ -7,8 +7,11 @@ import com.freya02.botcommands.api.components.Components;
 import dev.cybo.tickets.DevRoomTickets;
 import dev.cybo.tickets.objects.Ticket;
 import net.dv8tion.jda.api.EmbedBuilder;
+import net.dv8tion.jda.api.interactions.components.buttons.Button;
 
 import java.awt.*;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Timer;
 import java.util.TimerTask;
 
@@ -26,17 +29,33 @@ public class TicketOptionsCommand extends ApplicationCommand {
     )
     public void onSlashTicketOptions(GuildSlashEvent event) {
 
+        for (Ticket ticket : ticketBot.getStorage().getTicketList().values()) {
+            if (ticket.getAuthorId().equals(event.getUser().getId())) {
+                event.reply("You already have a ticket open!").setEphemeral(true).queue();
+                return;
+            }
+        }
+
         EmbedBuilder embedBuilder = new EmbedBuilder();
         embedBuilder.setTitle("Tickets");
         embedBuilder.setColor(Color.BLUE);
         embedBuilder.setDescription("What do you need help with?");
 
-        var replyCallbackAction = event.replyEmbeds(embedBuilder.build());
+        var replyCallbackAction = event.replyEmbeds(embedBuilder.build()).setEphemeral(true);
 
-        ticketBot.getConfig().getArray("ticketOptions").forEach(option -> {
+        List<Button> buttonList = new ArrayList<>();
+        for (Object option : ticketBot.getConfig().getArray("ticketOptions")) {
             String finalOption = option.toString();
-            replyCallbackAction.addActionRow(Components.primaryButton(e -> {
-                e.getMessage().delete().queue();
+            buttonList.add(Components.primaryButton(e -> {
+                event.getHook().deleteOriginal().queue();
+                e.getHook().deleteOriginal().queue();
+                for (Ticket ticket : ticketBot.getStorage().getTicketList().values()) {
+                    if (ticket.getAuthorId().equals(event.getUser().getId())) {
+                        e.reply("You already have a ticket open!").setEphemeral(true).queue();
+                        return;
+                    }
+                }
+
                 event.getGuild().createTextChannel("ticket-" + event.getUser().getName()).queue(textChannel -> {
 
                     textChannel.sendMessage("Hello, " + event.getUser().getAsMention() +
@@ -52,7 +71,10 @@ public class TicketOptionsCommand extends ApplicationCommand {
                                 @Override
                                 public void run() {
                                     if (ticketBot.getStorage().getAwaitingResponse().contains(textChannel.getId())) {
-                                        textChannel.delete().queue();
+                                        ticketBot.getStorage().getAwaitingResponse().remove(textChannel.getId());
+                                        if (event.getGuild().getTextChannelById(textChannel.getId()) != null) {
+                                            textChannel.delete().queue();
+                                        }
                                     }
                                 }
                             }, 180000);
@@ -61,7 +83,11 @@ public class TicketOptionsCommand extends ApplicationCommand {
 
                 });
             }).build(finalOption));
-        });
+        }
+
+        for (int i = 0; i < buttonList.size(); i += 5) {
+            replyCallbackAction.addActionRow(buttonList.subList(i, Math.min(i + 5, buttonList.size())));
+        }
 
         replyCallbackAction.queue();
     }
